+++ cloud/file-server/src/main/java/pers/acp/admin/file/FileServerApplication.java
@@ -1,0 +1,17 @@
+package pers.acp.admin.file;
+
+import org.springframework.boot.SpringApplication;
+import pers.acp.springcloud.common.annotation.AcpCloudAtomApplication;
+
+/**
+ * @author zhangbin by 09/04/2018 16:11
+ * @since JDK 11
+ */
+@AcpCloudAtomApplication
+public class FileServerApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FileServerApplication.class, args);
+    }
+
+}