+++ cloud/file-server/src/main/java/pers/acp/admin/file/constant/FileExpression.java
@@ -1,0 +1,10 @@
+package pers.acp.admin.file.constant;
+
+import pers.acp.admin.common.permission.BaseExpression;
+
+/**
+ * @author zhang by 01/02/2019
+ * @since JDK 11
+ */
+public interface FileExpression extends BaseExpression {
+}