+++ cloud/file-server/src/main/java/pers/acp/admin/file/constant/FileParams.java
@@ -1,0 +1,13 @@
+package pers.acp.admin.file.constant;
+
+import java.io.File;
+
+/**
+ * @author zhang by 16/04/2019
+ * @since JDK 11
+ */
+public interface FileParams {
+
+    String upLoadPath = File.separator + "file" + File.separator + "upload";
+
+}