+++ cloud/oauth-server/src/main/java/pers/acp/admin/oauth/controller/api/GateWayRouteController.java
@@ -10,7 +10,7 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import pers.acp.admin.common.base.BaseController;
-import pers.acp.admin.common.constant.path.oauth.OauthApi;
+import pers.acp.admin.oauth.constant.OauthApi;
 import pers.acp.admin.common.permission.BaseExpression;
 import pers.acp.admin.common.vo.InfoVO;
 import pers.acp.admin.oauth.domain.GateWayRouteDomain;